Move the FEATURES object out to the build only
